Added initial records for all but stmts and exprs.
The usage of each class should be fairly self-evident;
DecafClass for classes, DecafMethod for methods, etc.
I left out records for statements and expressions because I'm not
really sure what's the best way to represent them. Methods,
fields, etc. are very similar to tables, so I felt that classes would
be a good representation. Statements and expressions are more tree-like
so I don't feel that classes are a good match for them. Haven't thought
of a solution for them yet either.

There a few things to note that may not be readily apparent:
1) the global tables (class_table, method_table, constructor_table, and
field_table) are stored as class variables inside the respective
classes. They should not need to be manually populated, as the
constructors will that already.
2) DecafType instantiation is done through static methods. If you wnat
to create a new int type, call DecafType.int(). For a new class type,
call e.g. DecafType.user_defined('HelloWorld').

Not much else to say except that these classes are intended to be
flexible and should be changed if needed. For example, I'm not sure
if we'll know the var_table at Constructor or Method instantiation, so
that may have to get removed from the __init__ method and turned into an
add_var(self, var) instance method.