Added method declarations and block scoping.
The AST properly recognizes method declarations. See
fields_methods.decaf for a test.

I added support for block scoping of variables. I added some embedded
actions to the p_block and p_block_error rules (lines 182-196 of
decafparser.py). At the start of a block, a new block scope is pushed
onto a scope stack. At tne end of a block, it is popped off. Within a
method, there will always be at least two blocks within the scope.
The first block will contain the formal parameters, the second block
will contain the local parameters of the top-most block. Afterwards,
blocks will be dynamically pushed and popped on and off the stack as
they come in scope.

NOTE: This differs slightly from the specification that states that
formal parameters are "treated the same way as local variables declared
in the outer-most block of method body." Because I separated them into
two separate blocks, we will have to perform some additional error
checking since "the names of formal parameters of a method must be
distinct from the names of variables defined in the top-most block."