Type instantiation fixes.
In p_var_id, the current type is passed an argument to Type.__init__.
This causes Type to wrap all types in an extra level of user(_). This
causes types like user(int) to appear. The fix is just to detext when
a Type is passed in and to copy that instead of wrapping around it.

In p_type_id, the class object itself instead of the name is passed in.
The fix is just to grab the name field (baseclass.name) and pass that
to Type.__init__.