All statements are now recursively checked.
Previously, because of the order of the error checking, later parts
of compound statements would not be error checked if earlier parts
failed. For example, if the condition expression of an if statement
failed, then the 'then' and 'else' parts would never get checked. This
does all of the recursive checking up front. More work but more
complete.